feat: delete a stream
